backend/requirements.txt
numpy
matplotlib
pandas
scikit-learn
nltk
umap-learn
sentence_transformers
keybert
uvicorn
fastapi[standard]
python-multipart
httpx 
fitz
python-docx